Complete America Bitcoin ATM rebrand across app surfaces

+++ ios/Shared/LatestTransaction.swift
@@ -10,5 +10,5 @@ import Foundation
 
 struct LatestTransaction {
   let isUnconfirmed: Bool?
-  let epochValue: Int?
+  let epochValue: Int64?
 }
